% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_glm2.R
\name{fast_glm2}
\alias{fast_glm2}
\title{Run many regression models quickly, for many response variables.}
\usage{
fast_glm2(
  response = NULL,
  response_name = NULL,
  Xvar = NULL,
  dataset = NULL,
  adj_vars = NULL,
  family = "gaussian",
  exponentiate = FALSE,
  p_adjust_meth = "fdr",
  robust = FALSE,
  sort = TRUE,
  decimals = 3
)
}
\arguments{
\item{response}{Character vector of variable names to use as response of interest in separate regression models.}

\item{response_name}{Label to display for each response of interest}

\item{Xvar}{Main predictor to be used in each regression.}

\item{dataset}{Dataset in which these variables exist.}

\item{adj_vars}{Character vector of adjustment covariate names.}

\item{family}{Family for regression model, e.g. 'gaussian', 'binomial', 'poisson'.}

\item{exponentiate}{Whether or not to exponentiate the estimate from the model (e.g. for logistic or Poisson regression).}

\item{p_adjust_meth}{Method by which to provide adjustment of P-values.}

\item{robust}{If TRUE, returns robust standard error estimates.}

\item{sort}{If TRUE, sorts by p-value.}

\item{decimals}{Specifies the number of decimals to which to round estimates.}
}
\description{
Performs any type of cross-sectional regression analysis quickly for many response variables.
}
