% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_analysis_CCSR.R
\name{phewas_analysis_CCSR}
\alias{phewas_analysis_CCSR}
\title{Determine the association of a variable of interest by running Cox regression on time to event information and logistic regression on prevalence at baseline.
Generates PheWAS plots to assess the strenth of associations}
\usage{
phewas_analysis_CCSR(
  cohort_df,
  outcome_df = CCSR_outcomes,
  target,
  adj_vars = c("age", "sex"),
  save_as = "CCSR",
  n_event_cutoff = 10,
  exclude_categories = c("Congenital malformations", "Injury and poisoning consequences",
    "Health status factors", "Pregnancy and childbirth", "Perinatal conditions",
    "Unacceptable principal diagnosis"),
  exclude_history = TRUE,
  return_tables = TRUE,
  annotation_size = 5,
  skip_colors = NULL
)
}
\arguments{
\item{cohort_df}{A dataframe including participant ID and adjustment columns}

\item{outcome_df}{Result from CCSR_tte function}

\item{target}{Analysis variable column name}

\item{adj_vars}{List of variables to adjust models for}

\item{save_as}{Prefix for saved plot names}

\item{n_event_cutoff}{Diseases with fewer events are omitted}

\item{exclude_categories}{CCSR categories to omit}

\item{exclude_history}{If TRUE, only considers events with no history before baseline (default)}

\item{skip_colors}{Omit colors from the palette, useful to sync plots between cohorts}
}
\value{
Returns two dataframes with regression results and generates PheWAS plots
}
\description{
Determine the association of a variable of interest by running Cox regression on time to event information and logistic regression on prevalence at baseline.
Generates PheWAS plots to assess the strenth of associations
}
