% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_analysis_phecodes.R
\name{phewas_analysis_phecodes}
\alias{phewas_analysis_phecodes}
\title{Determine the association of a variable of interest by running Cox regression on time to event information and logistic regression on prevalence at baseline.
Generates PheWAS plots to assess the strenth of associations}
\usage{
phewas_analysis_phecodes(
  cohort_df = ghs_eps,
  outcome_df = phecode_outcomes,
  adj_vars = c("age", "sex"),
  exclude_categories = c("congenital anomalies", "injuries & poisonings",
    "pregnancy complications", "symptoms", "other"),
  save_as = "phecode",
  n_event_cutoff = 10,
  target = target.scaled,
  exclude_history = TRUE,
  add_events = NULL,
  add_events_name = NULL,
  add_events_category = NULL,
  return_tables = FALSE,
  annotation_size = 5,
  skip_colors = NULL
)
}
\arguments{
\item{cohort_df}{A dataframe including participant ID and adjustment columns}

\item{outcome_df}{Result from phecode_tte function}

\item{adj_vars}{List of variables to adjust models for}

\item{exclude_categories}{Phecode categories to omit}

\item{save_as}{Prefix for saved plot names}

\item{n_event_cutoff}{Diseases with fewer events are omitted}

\item{target}{Analysis variable column name}

\item{exclude_history}{If TRUE, only considers events with no history before baseline (default)}

\item{add_events}{Add a user-generated event to the analysis. Needs to exist in the outcome_df or cohort_df dataframe}

\item{skip_colors}{Omit colors from the palette, useful to sync plots between cohorts}
}
\value{
Returns two dataframes with regression results and generates PheWAS plots
}
\description{
Determine the association of a variable of interest by running Cox regression on time to event information and logistic regression on prevalence at baseline.
Generates PheWAS plots to assess the strenth of associations
}
