% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_event_CCSR.R
\name{time_to_event_CCSR}
\alias{time_to_event_CCSR}
\title{Map temporal ICD-10 codes into outcomes for Cox regression using the CCSR convention}
\usage{
time_to_event_CCSR(
  dataframe_ID_BL,
  icd_table_long,
  tbl_mortality_contact,
  censor_time = 5,
  one_to_one_mapping = TRUE,
  save_table = FALSE
)
}
\arguments{
\item{dataframe_ID_BL}{A dataframe with a participant ID column. Dataframe gets returned with ICD-10 information is added to it}

\item{icd_table_long}{A long table with ICD-10 information, including the ID, ICD-10 code and date on which it was diagnosed}

\item{tbl_mortality_contact}{A dataframe containing participant ID, baseline visit date and last contact/death date columns.}

\item{censor_time}{Time at which events are censored}

\item{one_to_one_mapping}{If false, will map all CCSR-ICD-10 pairs instead of the single default CCSR-ICD-10 pair}

\item{save_table}{Decide if the dataframe is written into the current directory}
}
\value{
Returns a dataframe, with all outcomes (event yes/no + time to event/censoring) and prevalences at baseline accoding to CCSR convention
}
\description{
Map temporal ICD-10 codes into outcomes for Cox regression using the CCSR convention
}
