% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_event_phecode.R
\name{time_to_event_phecode}
\alias{time_to_event_phecode}
\title{Map temporal ICD-10 codes into outcomes for Cox regression using the phecode convention}
\usage{
time_to_event_phecode(
  dataframe_ID_BL,
  icd_table_long,
  tbl_mortality_contact,
  censor_time = 5,
  save_table = FALSE
)
}
\arguments{
\item{dataframe_ID_BL}{A dataframe with a participant ID column. Dataframe gets returned with ICD-10 information is added to it}

\item{icd_table_long}{A long table with ICD-10 information, including the ID, ICD-10 code and date on which it was diagnosed}

\item{tbl_mortality_contact}{A dataframe containing participant ID, baseline visit date and last contact/death date columns.}

\item{censor_time}{Time at which events are censored}

\item{save_table}{Decide if the dataframe is written into the current directory}
}
\value{
Returns a dataframe, with all outcomes (event yes/no + time to event/censoring) and prevalences at baseline accoding to phecode convention
}
\description{
Map temporal ICD-10 codes into outcomes for Cox regression using the phecode convention
}
